+++ src/application/unit-of-work/unit-of-work.test.ts
@@ -141,6 +141,7 @@ function createMockBus(): EventBus<TestEvent> & { published: TestEvent[][] } {
 		subscribe: () => () => {},
 		subscribeAll: () => () => {},
 		once: () => new Promise(() => {}),
+		close: () => {},
 	};
 }
 
@@ -1362,6 +1363,7 @@ describe("UnitOfWork", () => {
 				subscribe: () => () => {},
 				subscribeAll: () => () => {},
 				once: () => new Promise(() => {}),
+				close: () => {},
 			};
 			const event = testEvent("o-1");
 			const agg = createMockAggregate("o-1", [event]);